# ⚡️ perf(ctx.Range): reduce allocations