## Description

`strings.Split` was causing extra allocations where `strings.Cut` can suffice. Also switch from `strconv.Atoi` because it causes an allocation when parsing a non-integer, which is common for Ranges.

The remaining allocation (see benchmarks) comes from the returned slice.

## Checklist:
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] New and existing unit tests pass locally with my changes
- [x] I tried to make my code as fast as possible with as few allocations as possible
- [x] For new code I have written benchmarks so that they can be analyzed and improved

## Bench:
```
                            │  master.txt  │             branch.txt              │
                            │    sec/op    │   sec/op     vs base                │
_Ctx_Range/bytes=-700-8       197.80n ± 0%   65.06n ± 0%  -67.11% (p=0.000 n=20)
_Ctx_Range/bytes=500--8       198.10n ± 0%   63.70n ± 0%  -67.84% (p=0.000 n=20)
_Ctx_Range/bytes=500-1000-8   162.30n ± 0%   64.76n ± 0%  -60.10% (p=0.000 n=20)
_Ctx_Range/bytes=500-700-8    161.75n ± 0%   64.48n ± 0%  -60.14% (p=0.000 n=20)
geomean                        179.1n        64.50n       -63.98%

                            │ master.txt  │             branch.txt             │
                            │    B/op     │    B/op     vs base                │
_Ctx_Range/bytes=-700-8       144.00 ± 0%   16.00 ± 0%  -88.89% (p=0.000 n=20)
_Ctx_Range/bytes=500--8       144.00 ± 0%   16.00 ± 0%  -88.89% (p=0.000 n=20)
_Ctx_Range/bytes=500-1000-8    96.00 ± 0%   16.00 ± 0%  -83.33% (p=0.000 n=20)
_Ctx_Range/bytes=500-700-8     96.00 ± 0%   16.00 ± 0%  -83.33% (p=0.000 n=20)
geomean                        117.6        16.00       -86.39%

                            │ master.txt │             branch.txt             │
                            │ allocs/op  │ allocs/op   vs base                │
_Ctx_Range/bytes=-700-8       5.000 ± 0%   1.000 ± 0%  -80.00% (p=0.000 n=20)
_Ctx_Range/bytes=500--8       5.000 ± 0%   1.000 ± 0%  -80.00% (p=0.000 n=20)
_Ctx_Range/bytes=500-1000-8   4.000 ± 0%   1.000 ± 0%  -75.00% (p=0.000 n=20)
_Ctx_Range/bytes=500-700-8    4.000 ± 0%   1.000 ± 0%  -75.00% (p=0.000 n=20)
geomean                       4.472        1.000       -77.64%
```