@ReneWerner87 2023-11-06
thanks for the optimization

unfortunately we can't do it like this
because of
![image](https://github.com/gofiber/fiber/assets/7063188/17ee7eb6-2a71-443a-9590-4509482ebcc5)
![image](https://github.com/gofiber/fiber/assets/7063188/d2e7596b-798c-4084-909b-a194357a5425)

maybe you can do the exchange of the cut function only for the golang version from 1.18 ? here an example
https://github.com/gofiber/fiber/blob/master/utils/convert_s2b_new.go#L1
https://github.com/gofiber/fiber/blob/master/utils/convert_s2b_old.go#L1