@nickajacks1 2023-11-06
Sure, I'll try out strings.IndexByte