@gaby 2023-11-07
@nickajacks1 

```
Error: ./ctx.go:1398:26: undefined: strings.Cut
Error: ./ctx.go:1410:40: undefined: strings.Cut
note: module requires Go 1.20
```