@nickajacks1 2023-11-10
Removed strings.Cut, added more tests, updated Benchmark to include a case with multiple ranges.
New benchmarks:
```
                                  │  master.txt  │             branch.txt              │
                                  │    sec/op    │   sec/op     vs base                │
_Ctx_Range/bytes=-700-8             198.20n ± 1%   55.46n ± 1%  -72.02% (p=0.000 n=20)
_Ctx_Range/bytes=500--8             197.50n ± 0%   55.08n ± 0%  -72.11% (p=0.000 n=20)
_Ctx_Range/bytes=500-1000-8         160.60n ± 0%   55.93n ± 1%  -65.18% (p=0.000 n=20)
_Ctx_Range/bytes=0-700,800-1000-8   245.10n ± 0%   93.25n ± 0%  -61.95% (p=0.000 n=20)
geomean                              198.1n        63.18n       -68.11%

                                  │ master.txt  │             branch.txt             │
                                  │    B/op     │    B/op     vs base                │
_Ctx_Range/bytes=-700-8             144.00 ± 0%   16.00 ± 0%  -88.89% (p=0.000 n=20)
_Ctx_Range/bytes=500--8             144.00 ± 0%   16.00 ± 0%  -88.89% (p=0.000 n=20)
_Ctx_Range/bytes=500-1000-8          96.00 ± 0%   16.00 ± 0%  -83.33% (p=0.000 n=20)
_Ctx_Range/bytes=0-700,800-1000-8   176.00 ± 0%   48.00 ± 0%  -72.73% (p=0.000 n=20)
geomean                              136.8        21.06       -84.61%

                                  │ master.txt │             branch.txt             │
                                  │ allocs/op  │ allocs/op   vs base                │
_Ctx_Range/bytes=-700-8             5.000 ± 0%   1.000 ± 0%  -80.00% (p=0.000 n=20)
_Ctx_Range/bytes=500--8             5.000 ± 0%   1.000 ± 0%  -80.00% (p=0.000 n=20)
_Ctx_Range/bytes=500-1000-8         4.000 ± 0%   1.000 ± 0%  -75.00% (p=0.000 n=20)
_Ctx_Range/bytes=0-700,800-1000-8   6.000 ± 0%   2.000 ± 0%  -66.67% (p=0.000 n=20)
geomean                             4.949        1.189       -75.97%
```